requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0

qdrant-client>=1.7.0
sentence-transformers>=2.2.2
python-dotenv>=1.0.0

markdown>=3.5.1
beautifulsoup4>=4.12.2

numpy>=1.24.0
